Integrate contingency plotting
